lib: pdump is not supported on Windows
The current version of the pdump library was building on
Windows, but it was useless since the pdump utility was not being
built and Windows does not have multi-process support.

The new version of pdump with filtering now has dependency
on bpf. But bpf library is not available on Windows.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Acked-by: Dmitry Kozlyuk <dmitry.kozliuk@gmail.com>

Cc: Dmitry Kozlyuk <dmitry.kozliuk@gmail.com>
Cc: Narcisa Ana Maria Vasile <navasile@linux.microsoft.com>
Cc: Dmitry Malloy <dmitrym@microsoft.com>
Cc: Pallavi Kadam <pallavi.kadam@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>