mempool: accept user flags only
As reported by Dmitry, MEMPOOL_F_POOL_CREATED is a flag only manipulated
internally.
This flag is not supposed to be requested from an application and would
probably result in an incorrect behavior if an application did pass it.

Other internal flags may be introduced later.

Rework the check and export a mask of valid user flags for use in the
unit test.

Fixes: b240af8b10f9 ("mempool: enforce valid flags at creation")

Reported-by: Dmitry Kozlyuk <dkozlyuk@nvidia.com>
Signed-off-by: David Marchand <david.marchand@redhat.com>
Acked-by: Olivier Matz <olivier.matz@6wind.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>