ethdev: fix log level of Tx and Rx dummy functions
When stopping a port, the data path Tx and Rx burst functions should
be stopped firstly conventionally. Then the dummy functions are used
to replace the callback functions provided by the PMD.

When the application stops a port without or before stopping the data
path handling. The dummy functions may be invoked heavily and a lot
of logs in these dummy functions will result in a flood.

Debug level log should be enough instead of the error level.

Fixes: c87d435a4d79 ("ethdev: copy fast-path API into separate structure")
Cc: konstantin.ananyev@intel.com

Signed-off-by: Bing Zhao <bingz@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>