net/hns3: fix mailbox communication with HW
Mailbox is the communication mechanism between SW and HW. There exist two
approaches for SW to recongnize mailbox message from HW. One way is using
match_id, the other is to compare the message code. The two approaches are
independent and used in different scenarios.

But for the second approache, "next_to_use" should be updated and written
to HW register. If it not done, HW do not know the position SW steps, then,
the communication between SW and HW will turn to be failed.

Fixes: dbbbad23e380 ("net/hns3: fix VF handling LSC event in secondary process")
Cc: stable@dpdk.org

Signed-off-by: Min Hu (Connor) <humin29@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>