net/ice: fix flow redirect failure
When the switch flow rules are redirected, if rule is removed but not
added successfully, the rule addition in the next time will not succeed
because the rule's meta cannot be found.

This patch uses the saved flow rule's meta when the flow rule is added
again to make the addition succeed.

Fixes: 397b4b3c5095 ("net/ice: enable flow redirect on switch")
Cc: stable@dpdk.org

Signed-off-by: Dapeng Yu <dapengx.yu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>